requirements.txt
asgiref==3.7.2
Django==4.2.3
packaging==23.1
pdf2image==1.16.3
Pillow==10.0.0
pytesseract==0.3.10
sqlparse==0.4.4
